requirements.txt
cloudinary==1.41.0
Django==5.1.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
drf_yasg==1.21.8
numpy==2.2.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==11.0.0
pytesseract==0.3.13
python-dotenv==1.0.1
weasyprint==63.1
celery
redis
django-celery-results
django-cors-headers
pymupdf


#crear tablas en sqlite
#python manage.py makemigrations
#python manage.py migrate
#python manage.py createsuperuser

# superuser = User.objects.create_superuser(
#     dni="12345678Z",
#     fullname="Voltix",
#     email="adminvoltix@igorwker.com",
#     password="1234Aa@"
# )
# Y

#rellenar con data
#python manage.py add

#probar entrar como usuario